+++ kffi-objc/src/jvmMain/kotlin/org/graphiks/kffi/objc/classes/SCDisplay.kt
@@ -1,0 +1,69 @@
+@file:OptIn(org.graphiks.kffi.objc.PlatformAvailability::class)
+
+package org.graphiks.kffi.objc
+
+import java.lang.invoke.*
+import java.lang.foreign.*
+import java.lang.foreign.MemoryLayout.PathElement.*
+
+/**
+ * Kotlin/JVM wrapper for Objective-C class: SCDisplay
+ * Superclass: NSObject
+ */
+@PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+@PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+open class SCDisplay(override val ptr: MemorySegment) : NSObject(ptr) {
+    companion object {
+        private val _class: MemorySegment by lazy { LOOKUP.let { ObjCRuntime.getClass("SCDisplay") } }
+
+        @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+        @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+        @PlatformAvailability(platform = "all", unavailable = true)
+        fun new(): MemorySegment {
+            val sel = ObjCRuntime.sel("new")
+            return ObjCRuntime.msgSend(ValueLayout.ADDRESS, _class, sel) as MemorySegment
+        }
+
+    }
+
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+    @PlatformAvailability(platform = "all", unavailable = true)
+    open fun init(): MemorySegment {
+        val sel = ObjCRuntime.sel("init")
+        return ObjCRuntime.msgSend(ValueLayout.ADDRESS, ptr, sel) as MemorySegment
+    }
+
+    // @property displayID
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+    open fun displayID(): Int {
+        val sel = ObjCRuntime.sel("displayID")
+        return ObjCRuntime.msgSend(ValueLayout.JAVA_INT, ptr, sel) as Int
+    }
+
+    // @property width
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+    open fun width(): Long {
+        val sel = ObjCRuntime.sel("width")
+        return ObjCRuntime.msgSend(ValueLayout.JAVA_LONG, ptr, sel) as Long
+    }
+
+    // @property height
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+    open fun height(): Long {
+        val sel = ObjCRuntime.sel("height")
+        return ObjCRuntime.msgSend(ValueLayout.JAVA_LONG, ptr, sel) as Long
+    }
+
+    // @property frame
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 12, introducedMinor = 3, introducedSubminor = -1)
+    open fun frame(): CGRect {
+        val sel = ObjCRuntime.sel("frame")
+        return CGRect(ObjCRuntime.msgSendStruct(CGRect.layout, ptr, sel))
+    }
+
+}