+++ kffi-objc/src/jvmMain/kotlin/org/graphiks/kffi/objc/classes/SCShareableContentInfo.kt
@@ -1,0 +1,45 @@
+@file:OptIn(org.graphiks.kffi.objc.PlatformAvailability::class)
+
+package org.graphiks.kffi.objc
+
+import java.lang.invoke.*
+import java.lang.foreign.*
+import java.lang.foreign.MemoryLayout.PathElement.*
+
+/**
+ * Kotlin/JVM wrapper for Objective-C class: SCShareableContentInfo
+ * Superclass: NSObject
+ */
+@PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+@PlatformAvailability(platform = "macos", introducedMajor = 14, introducedMinor = 0, introducedSubminor = -1)
+open class SCShareableContentInfo(override val ptr: MemorySegment) : NSObject(ptr) {
+    companion object {
+        private val _class: MemorySegment by lazy { LOOKUP.let { ObjCRuntime.getClass("SCShareableContentInfo") } }
+
+    }
+
+    // @property style
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 14, introducedMinor = 0, introducedSubminor = -1)
+    open fun style(): SCShareableContentStyle {
+        val sel = ObjCRuntime.sel("style")
+        return SCShareableContentStyle(ObjCRuntime.msgSend(ValueLayout.JAVA_LONG, ptr, sel) as Long)
+    }
+
+    // @property pointPixelScale
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 14, introducedMinor = 0, introducedSubminor = -1)
+    open fun pointPixelScale(): Float {
+        val sel = ObjCRuntime.sel("pointPixelScale")
+        return ObjCRuntime.msgSend(ValueLayout.JAVA_FLOAT, ptr, sel) as Float
+    }
+
+    // @property contentRect
+    @PlatformAvailability(platform = "maccatalyst", introducedMajor = 18, introducedMinor = 2, introducedSubminor = -1)
+    @PlatformAvailability(platform = "macos", introducedMajor = 14, introducedMinor = 0, introducedSubminor = -1)
+    open fun contentRect(): CGRect {
+        val sel = ObjCRuntime.sel("contentRect")
+        return CGRect(ObjCRuntime.msgSendStruct(CGRect.layout, ptr, sel))
+    }
+
+}